requirements.txt
streamlit
langchain
langchain-openai
langchain-pinecone
pypdf
python-dotenv
pinecone-client